MovieProphecy
An application to predict the success/failure of recent and upcoming
movies by sentimentally analyzing tweets. Gathered tweets related to
the movies and performed polarity analysis in order to categorize them
as positive/negative/ neutral. Used the ratio of the positive to the
negative tweets to predict the success/failure of a movie.